+++ views/student-orgs/index.js
@@ -1,3 +1,3 @@
 // @flow
 export {StudentOrgsView} from './list'
-export {StudentOrgsDetailView} from './detail'
+export {StudentOrgsDetailView} from './detail-wrapper'
